# perf(parser): memoize parsed regular expressions