--- a/Cargo.toml
+++ b/Cargo.toml
@@ -19,6 +19,7 @@ regex = "1"
 serde = { version = "1.0", features = ["derive"], optional = true }
 serde_regex = { version = "1.1", optional = true }
 fnv = "1"
+memoize = "0.4.0"
 
 [lib]
 crate-type = ["lib", "cdylib", "staticlib"]